fix: stub in global object
