+++ packages/create-instance/create-instance.js
@@ -55,7 +55,7 @@ export default function createInstance(
   const stubComponentsObject = createStubsFromStubsObject(
     componentOptions.components,
     // $FlowIgnore
-    { ...options.stubs, transition: true },
+    options.stubs,
     _Vue
   )
