+++ packages/test-utils/src/config.js
@@ -1,5 +1,8 @@
 export default {
-  stubs: {},
+  stubs: {
+    transition: true,
+    'transition-group': true
+  },
   mocks: {},
   methods: {},
   provide: {},
